+++ doc/src/changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Upcoming
+
+**Features**:
+
+- `GARDEN_CMD_VERBOSE` and `GARDEN_CMD_QUIET` are now updated when using
+`garden <command> -v ...` and `garden cmd <command> -v ...`.
+Previously, these variables were only set when `-v` was used against the `garden`
+command directly, before any sub-commands; e.g. `garden -v ...`.
+
+
 ## v1.9.0
 
 *Released 2024-10-11*